Global modification of the script
It now uses a hidden folder in the $HOME path.
There it manages the image with a folder system old/new.
Each terminal calls the script that moves the image in new to old.
Removes the old image & moves the downloaded one in new.

If no dowload is possible it loads a local image.